# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/policies": {
      "get": {
        "operationId": "PoliciesController_list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "clientId",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListPoliciesResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters."
          },
          "401": {
            "description": "Missing or invalid session."
          }
        },
        "security": [
          {
            "funes_vault_session": []
          }
        ],
        "summary": "List policies for the current user",
        "tags": ["policies"]
      },
      "post": {
        "operationId": "PoliciesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePolicyRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body."
          },
          "401": {
            "description": "Missing or invalid session."
          },
          "409": {
            "description": "This app already has permissions"
          }
        },
        "security": [
          {
            "funes_vault_session": []
          }
        ],
        "summary": "Create a client access policy",
        "tags": ["policies"]
      }
    },
    "/v1/policies/defaults/{clientId}": {
      "post": {
        "operationId": "PoliciesController_restore",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "This app has no first-party defaults"
          },
          "401": {
            "description": "Missing or invalid session."
          },
          "404": {
            "description": "App not found"
          },
          "409": {
            "description": "This app already has permissions"
          }
        },
        "security": [
          {
            "funes_vault_session": []
          }
        ],
        "summary": "Restore default permissions on an existing first-party app",
        "tags": ["policies"]
      }
    },
    "/v1/policies/{id}": {
      "patch": {
        "operationId": "PoliciesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePolicyRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body."
          },
          "401": {
            "description": "Missing or invalid session."
          },
          "404": {
            "description": "Policy not found."
          }
        },
        "security": [
          {
            "funes_vault_session": []
          }
        ],
        "summary": "Update a client access policy",
        "tags": ["policies"]
      },
      "delete": {
        "operationId": "PoliciesController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session."
          },
          "404": {
            "description": "Policy not found."
          }
        },
        "security": [
          {
            "funes_vault_session": []
          }
        ],
        "summary": "Delete a client access policy",
        "tags": ["policies"]
      }
    },
...
  },
  "info": {
    "title": "Funes Vault API",
    "description": "Privacy-first personal memory API for users, LLM clients, and MCP integrations.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "funes_vault_session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "funes_vault_session",
        "description": "HTTP-only session cookie established by Google sign-in."
      },
      "client_bearer": {
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "type": "http",
        "description": "Client bearer token returned at client creation or rotation."
      }
    },
    "schemas": {
...
      "PaginationDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 25
          },
          "total": {
            "type": "number",
            "example": 3
          },
          "totalPages": {
            "type": "number",
            "example": 1
          }
        },
        "required": ["page", "limit", "total", "totalPages"]
      },
...
      "PolicyDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cmqvt0v580000xeg7jgy7v1u3"
          },
          "clientId": {
            "type": "string",
            "example": "cmqvt0v580000xeg7jgy7v1u3"
          },
          "clientName": {
            "type": "object",
            "example": "Local Coding Agent",
            "nullable": true
          },
          "allowedCategoryKeys": {
            "example": ["communication_style", "software_development"],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "deniedCategoryKeys": {
            "example": ["health", "finance"],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "maxSensitivity": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "LOW",
              "INTERNAL",
              "SENSITIVE",
              "RESTRICTED",
              "SECRET"
            ],
            "example": "INTERNAL"
          },
          "operations": {
            "type": "array",
            "example": ["READ"],
            "items": {
              "type": "string",
              "enum": ["READ", "SUGGEST", "WRITE", "SUMMARIZE", "EXPORT"]
            }
          },
          "requiresConfirmation": {
            "type": "boolean",
            "example": true
          },
          "expiresAt": {
            "type": "object",
            "example": null,
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "example": "2026-06-27T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-06-27T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "clientId",
          "clientName",
          "allowedCategoryKeys",
          "deniedCategoryKeys",
          "maxSensitivity",
          "operations",
          "requiresConfirmation",
          "expiresAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "ListPoliciesResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PolicyDto"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationDto"
          }
        },
        "required": ["items", "pagination"]
      },
      "CreatePolicyRequestDto": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "string",
            "example": "cmqvt0v580000xeg7jgy7v1u3"
          },
          "allowedCategoryKeys": {
            "example": ["communication_style", "software_development"],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "deniedCategoryKeys": {
            "example": ["health", "finance"],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "maxSensitivity": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "LOW",
              "INTERNAL",
              "SENSITIVE",
              "RESTRICTED",
              "SECRET"
            ],
            "example": "INTERNAL"
          },
          "operations": {
            "type": "array",
            "example": ["READ", "SUGGEST", "WRITE"],
            "items": {
              "type": "string",
              "enum": ["READ", "SUGGEST", "WRITE", "SUMMARIZE", "EXPORT"]
            }
          },
          "requiresConfirmation": {
            "type": "boolean",
            "example": true
          },
          "expiresAt": {
            "type": "object",
            "example": null,
            "nullable": true
          }
        },
        "required": ["clientId"]
      },
      "PolicyResponseDto": {
        "type": "object",
        "properties": {
          "policy": {
            "$ref": "#/components/schemas/PolicyDto"
          }
        },
        "required": ["policy"]
      },
      "UpdatePolicyRequestDto": {
        "type": "object",
        "properties": {
          "allowedCategoryKeys": {
            "example": ["communication_style", "software_development"],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "deniedCategoryKeys": {
            "example": ["health", "finance"],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "maxSensitivity": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "LOW",
              "INTERNAL",
              "SENSITIVE",
              "RESTRICTED",
              "SECRET"
            ],
            "example": "INTERNAL"
          },
          "operations": {
            "type": "array",
            "example": ["READ"],
            "items": {
              "type": "string",
              "enum": ["READ", "SUGGEST", "WRITE", "SUMMARIZE", "EXPORT"]
            }
          },
          "requiresConfirmation": {
            "type": "boolean",
            "example": false
          },
          "expiresAt": {
            "type": "object",
            "example": null,
            "nullable": true
          }
        }
      },
...
    }
  }
}